Full testbench complete
Testbench complete along with makefile. Type in "make" to build and then "./bin/testbench" to run. For test file, type in "test_binaries/tst.bin"

+++ src/father.hpp
@@ -71,7 +71,7 @@ int32_t read_file(memory &mem, std::fstream &infile)
 {
     char c;
     uint32_t t = 0;
-    unsigned long count = 0;
+    uint32_t count = 0;
 
     while (!infile.eof())
     {
